Minor update and cleaned code
- Implemented attack range for player and enemy
- Implemented toggled minimap
- Fixed double death menu bug on killed
- Standardised identifiers
- Standardised data types
- Removed redundant code, comments, and variables
- Redesigned bar filling scripts:
BarFilling.cs -> parent
HealthBarFilling.cs -> child
StaminaBarFilling.cs -> child

P.S: Replaced StaminaFilling.cs

+++ Assets/Scripts/SkillAttack.cs
@@ -34,7 +34,7 @@ public class SkillAttack : MonoBehaviour {
     }
 
     // a method to handle Skill A
-    public void SkillA(int atk)
+    public void SkillA(float atk)
     {
         hit = Physics2D.Raycast(transform.position, currDir, 10.0f, targetLayer);
